pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plcforge"
version = "1.0.0"
description = "Multi-vendor PLC programming application with password recovery"
readme = "README.md"
license = {text = "Proprietary - Authorized Use Only"}
requires-python = ">=3.10"
authors = [
    {name = "PLCForge Team"}
]
keywords = ["plc", "siemens", "allen-bradley", "omron", "delta", "automation", "industrial"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications :: Qt",
    "Intended Audience :: Manufacturing",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Interface Engine/Protocol Translator",
]

dependencies = [
    # Communication protocols
    "python-snap7>=1.3",
    "pycomm3>=1.2.0",
    "pymodbus>=3.5.0",

    # GUI
    "PyQt6>=6.5.0",
    "pyqtgraph>=0.13.0",

    # AI/LLM
    "openai>=1.0.0",
    "anthropic>=0.20.0",
    "chromadb>=0.4.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-anthropic>=0.1.0",

    # Cryptography
    "cryptography>=41.0.0",

    # File parsing
    "olefile>=0.46",
    "lxml>=4.9.0",

    # Utilities
    "pydantic>=2.0.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-qt>=4.2.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

gpu = [
    "cupy-cuda12x>=12.0.0",  # GPU acceleration for password cracking
]

[project.scripts]
plcforge = "plcforge.main:main"

[project.gui-scripts]
plcforge-gui = "plcforge.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["plcforge*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = [
    "E501",   # Line too long (handled by formatter)
    "F401",   # Unused imports (needed for conditional imports)
    "E722",   # Bare except (acceptable in cleanup code)
    "B904",   # Exception chaining (not always needed)
]

[tool.ruff.lint.per-file-ignores]
"plcforge/drivers/*" = ["F401"]  # Conditional imports for optional dependencies
"plcforge/recovery/*" = ["N801"]  # Allow S7_xxx naming convention

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
